requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 6.2.1
PyYAML == 5.3.1
h5py == 2.10.0
ipython == 6.0.0
matplotlib == 3.1.1
numpy == 1.17.3
scipy == 0.19.1
torch == 1.5.0
tqdm == 4.37.0
nibabel == 2.2.1
